Fix: Prevent buffer overflow in create_element
In the original code of create_element, the number of characters
to be copied from s to val using strncpy is passed with strlen(val).
Since val isn't initialized yet, it would return an incorrect value
and therefore might result in buffer overflow since strncpy would
pad the destination with zeros until a total number of strlen(val)
have been written to it.

Change-Id: I54eef06abdb547b6802d6e69366b813b58282340