Implement monotonic queue operations
Add implementation for q_ascend() and q_descend(). The objective of
q_ascend() is equivalent to making the queue non-increasing from the
right and q_descend() is equivalent to making the queue non-decreasing
from the right.

Created monotonic_from_right() helper function to make the queue
non-decreasing or non-increasing from the right to the left(last to
first). An enumeration type "Order" is added to serve as a parameter
for monotonic_from_right() to indicate which kind of operation it
should perform.

Change-Id: I12de3e5db430f3256fd556e0e4ff968a5155fe13