Implement merge sort for queue
Add implementation for q_sort(), which uses bottom-up merge sort
approach. The function would merge two adjacent lists by calling
merge_two_lists() from index 0 and with size 1, 2, 4, 8, ... . The
merged lists are not guaranteed to be circular, thus rebuild_list()
is added to re-link each node's pointer to previous node.

In this implementation, a stack with size "STACKSIZE"(10^5) is declared
to store all nodes in the queue to provide random access. It could be
optimized by using other approaches such as 2:1 balanced merges that is
used in linux kernel.

Change-Id: Id983962cebb7894822b8482855f2d8e768b78f39