Implement queue merge operation
Add implementation for q_merge(). The function first moves all nodes
from the second to the last queue to the first one. All headers of the
queues except the first one is initialized to point to itself. Then we
q_sort() to sort the queue and return its size. If the chain contains
only one queue, it is guaranteed that the first queue is sorted
already, so the size is returned directly without calling q_sort() to
improve efficiency.

Change-Id: Ia67d21ad3be2dc5f7750bb4ef2f731c78b43a549