Optimize merge sort with fewer stack memory used
In the original version of q_sort(), we needed O(n) space to store all
nodes in the list to perform merge sort. In this approach, nodes are
pushed to the stack one by one, and is merged when there are two sorted
list with the same size. This approach uses O(log(n)) space, which is
much better than the previous method.

A new type "stack_t" and helper function s_push() and s_pop is added
to simplify the operation for stack and improves readability. The
"STACKSIZE" is reduced from 100000 to 32.

Change-Id: I907c25f664b31f4e2368df243ada3c924bdea349