Fix implementation of dudect
In the paper "Dude, is my code constant time?", the author mentioned
to interleave the sequence of classes during the evaluation. This could
prevent interference due to any concurrent process that may be
executing in parallel. However, the current implementation did not
apply this rule.

In this commit, update_statistics() is changed to discard the first 10
measurements, which is the same as how "oreparaz/dudect"
implements.

Link: https://github.com/oreparaz/dudect
Link: https://eprint.iacr.org/2016/1123.pdf

Change-Id: I8d9a3165468e2001ef52b2c4dede3251c610539f