Fix dudect cropping implementation
While measuring time complexity with dudect, the author proposed to
interleave the sequence of classes during the evaluation, preventing
interference due to any concurrent process that may be executing in
parallel.

In the previous implementation, DROP_SIZE is defined to drop the first
and last 20 measurements that is made by measure(). However, the loop
was implemented incorrectly, no measurements were made to discard.
The implementation would only result in smaller measurement sizes with
biases.

In this commit, the issue is resolved by changing the range of the loop
index to [0, N_MEASURES) and only write the data of ticks if index is in
range [DROP_SIZE, N_MEASURES-DROP_SIZE). Additionally, VALID_MEASURES
is defined to get number of measures that isn't dropped easily.

Change-Id: I856a5835da7ad05516c93ffa4de73f3fef6ebdf3