Introduce percentile for dudect
Add "Cropping" post-processing to measurements. This is to prevent
inaccurate data caused by measurement artifacts, main process being
interrupted by the OS etc. This method is introduced in "dudect".

In this commit, we set the percentile to 0.9, this value isn't derived
through strict mathematical calculation. Further experiments should be
made in future commits.

Change-Id: Ib7e66d564403add061ac38b4c89454c6a8221d5e