Fix dudect cutoff percentile
Originally, the cutoff value is only calculated for the first set of
measurements and applied to the rest. We observe that the first set of
measurement might be highly biased after running this code on github
actions.

In this commit, we discard the first set of measurement for each test
and calculate cutoff value for each set of measurements.

Change-Id: I9dda7e6fc39182525da45e35d7e6dea7e422f0d7